requirements.txt
amqp==5.0.6
billiard==3.6.4.0
cachetools==4.2.2
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colab==1.13.5
cssselect==1.1.0
cycler==0.10.0
diazo==1.4.0
Django==1.7.11
django-haystack==2.4.1
django-hitcounter==0.1.1
django-revproxy==0.10.0
django-taggit==0.22.2
et-xmlfile==1.1.0
etiquetando==0.1
future==0.18.2
google-api-core==1.26.3
google-auth==1.30.0
google-cloud-language==2.0.0
googleapis-common-protos==1.53.0
grpcio==1.37.1
html2text==2020.1.16
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
kombu==5.0.2
libcst==0.3.18
lxml==4.6.3
matplotlib==3.4.1
mypy-extensions==0.4.3
nltk==3.6.2
numpy==1.20.2
oauthlib==3.1.0
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
prompt-toolkit==3.0.18
proto-plus==1.18.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.4.4
repoze.lru==0.7
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
stemming==1.0.1
textblob==0.15.3
tqdm==4.60.0
tweepy==3.10.0
typing-extensions==3.10.0.0
typing-inspect==0.6.0
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
Whoosh==2.7.4
